requirements.txt
gunicorn==20.0.4
absl-py 
alabaster==0.7.12
argh==0.26.2
argon2-cffi 
asn1crypto 
astor==0.8.1
astroid 
astropy 
astunparse==1.6.3
async-generator==1.10
atomicwrites==1.4.0
attrs 
autopep8 
Babel 
backcall==0.2.0
backports.functools-lru-cache 
backports.shutil-get-terminal-size==1.0.0
backports.tempfile==1.0
backports.weakref==1.0.post1
bcrypt 
beautifulsoup4 
bitarray 
bkcharts==0.2
bleach 
bokeh 
boto==2.49.0
Bottleneck==1.3.2
brotlipy==0.7.0
cachetools==4.1.1
certifi==2020.6.20
cffi 
chardet 
click==7.1.2
cloudpickle 
colorama 
comtypes==1.1.7
contextlib2==0.6.0.post1
cryptography 
cycler==0.10.0
Cython==0.29.14
cytoolz==0.11.0
dask 
decorator==4.4.2
defusedxml==0.6.0
diff-match-patch 
distributed 
docutils==0.16
entrypoints==0.3
et-xmlfile==1.0.1
fastcache==1.1.0
filelock==3.0.12
flake8 
Flask==1.1.1
fsspec 
future==0.18.2
gast==0.2.2
gensim==3.8.3
gevent==1.4.0
glob2==0.7
google-auth==1.23.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
greenlet 
grpcio 
h5py==2.10.0
HeapDict==1.0.1
html5lib 
idna==2.5 
imagecodecs
imageio 
imagesize==1.2.0
imbalanced-learn==0.7.0
imblearn==0.0
importlib-metadata 
iniconfig 
intervaltree 
ipykernel 
ipython
ipython-genutils 
ipywidgets 
isort 
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.14.1
Jinja2==2.11.2
joblib 
json5==0.9.5
jsonschema 
jupyter==1.0.0
jupyter-client 
jupyter-console 
jupyter-core 
jupyterlab==2.2.6
jupyterlab-pygments 
jupyterlab-server==1.1.5 
Keras==2.3.1
Keras-Applications 
Keras-Preprocessing==1.1.2
keyring 
kiwisolver 
lazy-object-proxy==1.4.3
libarchive-c==2.9
llvmlite==0.34.0
locket==0.2.0
lxml 
Markdown 
MarkupSafe 
matplotlib 
mccabe==0.6.1
mistune 
mock==4.0.2
more-itertools
mpmath==1.1.0
msgpack==1.0.0
multipledispatch==0.6.0
nbclient 
nbconvert 
nbformat 
nest-asyncio 
networkx 
nltk 
nose 
notebook 
numba==0.51.2
numexpr==2.7.1
numpy==1.17.0
numpydoc 
oauthlib==3.1.0
olefile==0.46
openpyxl 
opt-einsum==3.3.0
packaging 
pandas
pandocfilters 
paramiko 
parso==0.5.2
partd==1.1.0
path 
pathlib2 
pathtools==0.1.2
patsy==0.5.1
pep8==1.7.1
pexpect 
pickleshare 
Pillow==4.3.0
pkginfo==1.6.1
pluggy==0.13.1
ply==3.11
prometheus-client 
prompt-toolkit 
protobuf==3.13.0
psutil 
ptyprocess 
py 
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycosat==0.6.3
pycparser 
pycrypto==2.6.1
pycurl==7.43.0.6
pydocstyle 
pyerfa 
pyflakes==2.2.0
Pygments 
pylint 
PyNaCl 
pyOpenSSL
pyparsing==2.4.7
pyreadline==2.1
pyrsistent 
PySocks
python-dateutil==2.8.1
python-jsonrpc-server 
python-language-server==0.31.7
pytz 
PyWavelets 
pywin32-ctypes 
pywinpty==0.5.7
PyYAML==5.3.1
pyzmq==20.0.0
QDarkStyle==2.8.1
QtAwesome 
qtconsole 
QtPy==1.9.0
regex 
requests 
requests-oauthlib==1.3.0
rope 
rsa==4.6
ruamel-yaml==0.15.87
scikit-image==0.17.2
scikit-learn 
scipy
seaborn 
Send2Trash==1.5.0
simplegeneric==0.8.1
singledispatch 
six 
smart-open==4.0.1
snowballstemmer==2.0.0
sortedcollections==1.2.1
sortedcontainers 
soupsieve==2.0.1
Sphinx 
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sphinxcontrib-websupport 
spyder==4.0.1
spyder-kernels==1.8.1
SQLAlchemy 
statsmodels==0.12.1
sympy 
tables==3.6.1
tblib 
tensorboard==2.0.2
tensorboard-plugin-wit==1.7.0
tensorflow==2.0.1
tensorflow-cpu==2.3.1
tensorflow-estimator==2.0.1
termcolor==1.1.0
terminado==0.9.1
testpath==0.4.4
textblob==0.15.3
threadpoolctl 
tifffile 
toml 
toolz 
tornado 
tqdm 
traitlets 
tweepy==3.9.0
typed-ast==1.4.1
typing-extensions 
ujson==3.0.0
unicodecsv==0.14.1
urllib3 
watchdog 
wcwidth 
webencodings==0.5.1
Werkzeug==0.16.0
widgetsnbextension==3.5.1
win-inet-pton 
win-unicode-console==0.5
wincertstore==0.2
wrapt==1.11.2
xgboost==1.2.1
xlrd==1.2.0
XlsxWriter 
xlwt==1.3.0
xmltodict==0.12.0
yapf
zict==2.0.0
zipp 
zope.event==4.5.0
zope.interface 
